adjust proto (#104)
