+++ Cargo.toml
@@ -45,8 +45,13 @@ path = "examples/hello_world/server.rs"
 serde_json = "1.0"
 serde = "1.0"
 serde_derive = "1.0"
-grpcio-proto = { path = "proto" }
+grpcio-proto = { path = "proto", version = "0.1.2" }
 rand = "0.3"
+slog = "2.0"
+slog-async = "2.1"
+slog-stdlog = "3.0"
+slog-scope = "4.0"
+slog-term = "2.2"
 
 [profile.release]
 debug = true
