+++ examples/hello_world/server.rs
@@ -21,6 +21,9 @@ extern crate grpcio_proto;
 #[macro_use]
 extern crate log;
 
+#[path = "../log_util.rs"]
+mod log_util;
+
 use std::io::Read;
 use std::sync::Arc;
 use std::{io, thread};
@@ -31,7 +34,6 @@ use grpcio::{Environment, RpcContext, ServerBuilder, UnarySink};
 
 use grpcio_proto::example::helloworld::{HelloReply, HelloRequest};
 use grpcio_proto::example::helloworld_grpc::{self, Greeter};
-use grpcio_proto::util;
 
 #[derive(Clone)]
 struct GreeterService;
@@ -48,7 +50,7 @@ impl Greeter for GreeterService {
 }
 
 fn main() {
-    let _guard = util::init_log();
+    let _guard = log_util::init_log();
     let env = Arc::new(Environment::new(1));
     let service = helloworld_grpc::create_greeter(GreeterService);
     let mut server = ServerBuilder::new(env)
