+++ examples/route_guide/server.rs
@@ -25,6 +25,8 @@ extern crate serde_derive;
 extern crate serde_json;
 
 mod util;
+#[path = "../log_util.rs"]
+mod log_util;
 
 use std::sync::Arc;
 use std::io::Read;
@@ -137,7 +139,7 @@ impl RouteGuide for RouteGuideService {
 }
 
 fn main() {
-    let _guard = grpcio_proto::util::init_log();
+    let _guard = log_util::init_log();
     let env = Arc::new(Environment::new(2));
     let instance = RouteGuideService {
         data: Arc::new(load_db()),
